requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
pip==22.3.1
python-telegram-bot==13.15
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7.1
setuptools==66.1.1
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
Flask~=2.2.2
deta~=1.1.0